# Better ExternalName support