--- a/pilot/pkg/model/sidecar.go
+++ b/pilot/pkg/model/sidecar.go
@@ -674,13 +674,13 @@ func (ilw *IstioEgressListenerWrapper) selectServices(services []*Service, confi
 
 		// Check if there is an explicit import of form ns/* or ns/host
 		if importedHosts, nsFound := hostsByNamespace[configNamespace]; nsFound {
-			if svc := matchingService(importedHosts, s, ilw); svc != nil {
+			if svc := matchingAliasService(importedHosts, matchingService(importedHosts, s, ilw)); svc != nil {
 				importedServices = append(importedServices, svc)
 				continue
 			}
 		}
 		if wnsFound { // Check if there is an import of form */host or */*
-			if svc := matchingService(wildcardHosts, s, ilw); svc != nil {
+			if svc := matchingAliasService(wildcardHosts, matchingService(wildcardHosts, s, ilw)); svc != nil {
 				importedServices = append(importedServices, svc)
 			}
 		}
@@ -730,6 +730,28 @@ func matchingService(importedHosts hostClassification, service *Service, ilw *Is
 		}
 	}
 	return nil
+}
+
+// Return the original service or a trimmed service which has a subset of the ports in original service.
+func matchingAliasService(importedHosts []host.Name, service *Service) *Service {
+	if service == nil {
+		return nil
+	}
+	matched := make([]NamespacedHostname, 0, len(service.Attributes.Aliases))
+	for _, importedHost := range importedHosts {
+		for _, alias := range service.Attributes.Aliases {
+			// Check if the hostnames match per usual hostname matching rules
+			if alias.Hostname.SubsetOf(importedHost) {
+				matched = append(matched, alias)
+			}
+		}
+	}
+	if len(matched) == len(service.Attributes.Aliases) {
+		return service
+	}
+	service = service.DeepCopy()
+	service.Attributes.Aliases = matched
+	return service
 }
 
 // serviceMatchingListenerPort constructs service with listener port.