--- a/pilot/pkg/networking/core/v1alpha3/httproute.go
+++ b/pilot/pkg/networking/core/v1alpha3/httproute.go
@@ -388,6 +388,10 @@ func BuildSidecarOutboundVirtualHosts(node *model.Proxy, push *model.PushContext
 
 	servicesByName := make(map[host.Name]*model.Service)
 	for _, svc := range services {
+		if svc.Resolution == model.Alias {
+			// Will be handled by the service it is an alias for
+			continue
+		}
 		if listenerPort == 0 {
 			// Take all ports when listen port is 0 (http_proxy or uds)
 			// Expect virtualServices to resolve to right port
@@ -403,6 +407,8 @@ func BuildSidecarOutboundVirtualHosts(node *model.Proxy, push *model.PushContext
 					Namespace:       svc.Attributes.Namespace,
 					ServiceRegistry: svc.Attributes.ServiceRegistry,
 					Labels:          svc.Attributes.Labels,
+					Aliases:         svc.Attributes.Aliases,
+					K8sAttributes:   svc.Attributes.K8sAttributes,
 				},
 			}
 			if features.EnableDualStack {
@@ -614,10 +620,18 @@ func generateVirtualHostDomains(service *model.Service, listenerPort int, port i
 		// Indicate we do not need port, as we will set IgnorePortInHostMatching
 		port = portNoAppendPortSuffix
 	}
-	altHosts := GenerateAltVirtualHosts(string(service.Hostname), port, node.DNSDomain)
-	domains := make([]string, 0, 4+len(altHosts))
-	domains = appendDomainPort(domains, string(service.Hostname), port)
-	domains = append(domains, altHosts...)
+	domains := []string{}
+	allAltHosts := []string{}
+	all := []string{string(service.Hostname)}
+	for _, a := range service.Attributes.Aliases {
+		all = append(all, a.Hostname.String())
+	}
+	for _, s := range all {
+		altHosts := GenerateAltVirtualHosts(s, port, node.DNSDomain)
+		domains = appendDomainPort(domains, s, port)
+		domains = append(domains, altHosts...)
+		allAltHosts = append(allAltHosts, altHosts...)
+	}
 
 	if service.Resolution == model.Passthrough &&
 		service.Attributes.ServiceRegistry == provider.Kubernetes {
@@ -638,7 +652,7 @@ func generateVirtualHostDomains(service *model.Service, listenerPort int, port i
 		domains = appendDomainPort(domains, addr, port)
 	}
 
-	return domains, altHosts
+	return domains, allAltHosts
 }
 
 // appendDomainPort appends `domain` and `domain:port` to `domains`. The `domain:port` variant is skipped