--- a/pilot/pkg/serviceregistry/kube/controller/controller.go
+++ b/pilot/pkg/serviceregistry/kube/controller/controller.go
@@ -427,7 +427,7 @@ func (c *Controller) addOrUpdateService(curr *v1.Service, currConv *model.Servic
 	}
 
 	// For ExternalName, we need to update the EndpointIndex, as we will store endpoints just based on the Service.
-	if curr != nil && curr.Spec.Type == v1.ServiceTypeExternalName {
+	if !features.EnableExternalNameAlias && curr != nil && curr.Spec.Type == v1.ServiceTypeExternalName {
 		updateEDSCache = true
 	}
 
@@ -469,7 +469,9 @@ func (c *Controller) buildEndpointsForService(svc *model.Service, updateCache bo
 		fep := c.collectWorkloadInstanceEndpoints(svc)
 		endpoints = append(endpoints, fep...)
 	}
-	endpoints = append(endpoints, kube.ExternalNameEndpoints(svc)...)
+	if !features.EnableExternalNameAlias {
+		endpoints = append(endpoints, kube.ExternalNameEndpoints(svc)...)
+	}
 	return endpoints
 }