… pilot/pkg/model/push_context_test.go
… pilot/pkg/networking/core/v1alpha3/route/route_cache_test.go
… pilot/pkg/xds/bench_test.go
… pilot/pkg/xds/testdata/benchmarks/externalname.yaml
… pkg/test/framework/components/istio/ingress.go
… releasenotes/notes/external-name.yaml
… tests/integration/base.yaml
… tests/integration/pilot/common/routing.go
… tests/integration/pilot/common/traffic.go