Github won't let me use my old PR (https://github.com/istio/istio/pull/37365) so making a new one...
Fixes https://github.com/istio/istio/issues/37331

Example config:
```yaml
apiVersion: v1
kind: Service
metadata:
  name: b-ext
  namespace: echo
spec:
  ports:
  - name: tcp
    port: 80
    targetPort: 18080
  type: ExternalName
  externalName: b.echo.svc.cluster.local
```

Kubernetes behavior: `b-ext` creates a CNAME record for `b`. Port is not relevant.

Istio (today): Create a DNS cluster for each port, which ends up resolving to Cluster IP. Match on Host header (http), SNI (tls), or just port (TCP) (meaning it captures all traffic on that port).

Issues:
* Port matters in Istio but not Kubernetes
* We get minimal functionality because we only have the cluster IP. No load balancing, auto mtls, etc
* TCP doesn't work well since there is no VIP - instead we capture ALL traffic on the configured port 

This PR changes the implementation to more closely align with Kubernetes.

ExternalName becomes a new "Alias" service type internally. An alias simply adds extra matching logic to _existing_ Services.  For example, above we would add a match for `b-ext` everywhere we match for `b`. This works for HTTP and TLS; for TCP it is already not needed as it works today (envoy will see traffic to `b` and `b-ext` as the same). If there was no Service/ServiceEntry defined at all for the target `externalName`, then we would end up just passing through.

Note that port from ExternalName is not used at all, just like in Kubernetes.

If we have a `destination` in VirtualService referring to an ExternalName service explicitly, this will now no longer work unless they have also defined the referenced `externalName` as a concrete service. For example:
```
apiVersion: v1
kind: Service
metadata:
  name: google
spec:
  type: ExternalName
  externalName: google.com
```
If I forward to this in a VS, it will work IFF I have a ServiceEntry defined for google.com. As such, it is generally not recommended to do this (you could just point to google.com directly), but can be useful when use third party Ingress configs, etc. Other Ingress implementations have extremely inconsistent support for ExternalName -- I couldn't find two implementations that behaved the same way -- so there isn't much of a standard to follow here.

One major difference in functionality with this change is that the ExternalName service is now just an alias, not its own "thing" (Envoy cluster). This means we cannot apply different rules for it. However, this can be done with standard DR subsets.

TODO:
* ~Add tests. Critical ones are Sidecar scoping~
* ~Add feature flag to retain old behavior~
* ~Add release note~
* How should VS behave? if I bind to the alias service, does it apply or not?