@istio-testing 2023-08-04
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
unit-tests-arm64_istio | 645013a3d3b1a351713ce7e8117f1b4277566ffe | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/46332/unit-tests-arm64_istio/1687576000465997824) | true | `/test unit-tests-arm64`
unit-tests_istio | 645013a3d3b1a351713ce7e8117f1b4277566ffe | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/46332/unit-tests_istio/1687576000424054784) | true | `/test unit-tests`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->