@howardjohn 2023-10-26
@hzxuzhonghu added benchmarks. 100 and 5k ExternalName services tested:

```
name                                    old time/op        new time/op        delta
InitPushContext/externalname-100-8             108µs ±14%         135µs ±15%  +24.81%  (p=0.000 n=8+10)
InitPushContext/externalname-5000-8           12.9ms ±10%        11.4ms ± 9%  -12.06%  (p=0.000 n=10+9)
RouteGeneration/externalname-100-8             400µs ± 2%           7µs ± 2%  -98.34%  (p=0.000 n=9+8)
RouteGeneration/externalname-5000-8           32.6ms ±12%         0.0ms ±15%  -99.94%  (p=0.000 n=10+9)
ClusterGeneration/externalname-100-8          1.22ms ± 3%        0.03ms ±32%  -97.41%  (p=0.000 n=9+10)
ClusterGeneration/externalname-5000-8         66.0ms ± 3%         0.0ms ±16%  -99.93%  (p=0.000 n=9+10)
ListenerGeneration/externalname-100-8          105µs ± 2%         138µs ±24%  +31.60%  (p=0.000 n=9+10)
ListenerGeneration/externalname-5000-8        1.39ms ± 7%        0.75ms ±17%  -45.88%  (p=0.000 n=9+9)

name                                    old alloc/op       new alloc/op       delta
InitPushContext/externalname-100-8            68.8kB ± 1%       114.1kB ± 0%  +65.78%  (p=0.000 n=10+10)
InitPushContext/externalname-5000-8           8.02MB ± 5%        8.21MB ± 5%     ~     (p=0.182 n=10+9)
RouteGeneration/externalname-100-8             368kB ± 0%           5kB ± 0%  -98.67%  (p=0.000 n=10+10)
RouteGeneration/externalname-5000-8           17.3MB ± 0%         0.0MB ± 0%  -99.97%  (p=0.000 n=10+10)
ClusterGeneration/externalname-100-8           574kB ± 0%          16kB ± 0%  -97.21%  (p=0.000 n=9+10)
ClusterGeneration/externalname-5000-8         28.5MB ± 0%         0.0MB ± 0%  -99.94%  (p=0.000 n=10+8)
ListenerGeneration/externalname-100-8         64.2kB ± 0%        64.2kB ± 0%   +0.00%  (p=0.009 n=9+9)
ListenerGeneration/externalname-5000-8         143kB ± 0%         143kB ± 0%     ~     (p=0.123 n=9+10)

name                                    old allocs/op      new allocs/op      delta
InitPushContext/externalname-100-8               579 ± 0%           401 ± 0%  -30.74%  (p=0.000 n=10+10)
InitPushContext/externalname-5000-8            25.2k ± 0%         15.4k ± 0%  -39.16%  (p=0.000 n=10+9)
RouteGeneration/externalname-100-8             2.80k ± 0%         0.05k ± 0%  -98.21%  (p=0.000 n=10+10)
RouteGeneration/externalname-5000-8             135k ± 0%            0k ± 0%  -99.96%  (p=0.000 n=10+10)
ClusterGeneration/externalname-100-8           9.23k ± 0%         0.22k ± 0%  -97.64%  (p=0.000 n=10+10)
ClusterGeneration/externalname-5000-8           457k ± 0%            0k ± 0%  -99.95%  (p=0.000 n=10+10)
ListenerGeneration/externalname-100-8            630 ± 0%           630 ± 0%     ~     (all equal)
ListenerGeneration/externalname-5000-8         5.53k ± 0%         5.53k ± 0%     ~     (all equal)

name                                    old kb/msg         new kb/msg         delta
RouteGeneration/externalname-100-8              41.0 ± 0%           0.4 ± 0%  -99.00%  (p=0.000 n=10+10)
RouteGeneration/externalname-5000-8            2.07k ± 0%         0.00k ± 0%  -99.98%  (p=0.000 n=10+10)
ClusterGeneration/externalname-100-8            59.2 ± 0%           1.8 ± 0%  -97.01%  (p=0.000 n=10+10)
ClusterGeneration/externalname-5000-8          2.93k ± 0%         0.00k ± 0%  -99.94%  (p=0.000 n=10+10)
ListenerGeneration/externalname-100-8           14.7 ± 0%          14.7 ± 0%     ~     (all equal)
ListenerGeneration/externalname-5000-8          14.7 ± 0%          14.7 ± 0%     ~     (all equal)

name                                    old resources/msg  new resources/msg  delta
RouteGeneration/externalname-100-8              1.00 ± 0%          1.00 ± 0%     ~     (all equal)
RouteGeneration/externalname-5000-8             1.00 ± 0%          1.00 ± 0%     ~     (all equal)
ClusterGeneration/externalname-100-8             105 ± 0%             5 ± 0%  -95.24%  (p=0.000 n=10+10)
ClusterGeneration/externalname-5000-8          5.00k ± 0%         0.01k ± 0%  -99.90%  (p=0.000 n=10+10)
ListenerGeneration/externalname-100-8           3.00 ± 0%          3.00 ± 0%     ~     (all equal)
ListenerGeneration/externalname-5000-8          3.00 ± 0%          3.00 ± 0%     ~     (all equal)
```

So in init PushContext (which I think is where your concerns were) there is a _slight_ increase for a small number of ExternalName. if we start to have a lot, its actually better. 

Here is zero externalNames (no diff):

```
name                    old time/op    new time/op    delta
InitPushContext/http-8     142µs ±13%     141µs ±12%   ~     (p=0.931 n=9+9)

name                    old alloc/op   new alloc/op   delta
InitPushContext/http-8    89.0kB ± 0%    89.1kB ± 1%   ~     (p=0.188 n=9+10)

name                    old allocs/op  new allocs/op  delta
InitPushContext/http-8       671 ± 0%       671 ± 0%   ~     (all equal)
```

The big difference is in XDS gen, which is incredibly more efficient in this PR (since we generate way way less XDS)

So overall I think this will be a big improvement in performance for most cases